Readding amandaMeasures.R to Main branch
I had deleted the previous version of amandaMeasures.R due to conflicting with and attempted merger between this repository and the original under control of Jon.  This is the older version of this file as I will go through the code with Jon so that we are running similar versions.