Updated Master Dataset
Went through ArchaicUng_Upload_MasterFile to clean up redundant entries (same paper and number repeated) and fill in column so that redundant entries form different papers (i.e. one paper reported measures of another) so that  the paper with most measures will be used, or defaulted to the original paper.