Create src_evanproposal.R
Adding src file that holds Evan's own functions and edits to Jon's functions.  This file is a testbed for things that can be implemented into the code of other src files pending approval by Jon.  Some code may be deprecated and needs to be be cleaned at a later time.