Removed Depricated Functions and Fixes to main_ecologyAnalysis
I went and removed a lot of deprecated functions from evanprposal.R.  Many of these functions were made to take a lot of Jon old code and place it into a function instead of running lines themselves.  Most of this was deprecated.  Other function removed include those that are now redundant or not needed (e.g., datacoveragev1 and v2).