Fixed getCurrentHigherTaxonomy function
reworked the function so that it is working off of the PBDB for setting accepted name, genus, and species.  Fixed some issues related to how previous code was not always overwriting things (often due to indexing not working correctly; resulting in NA's). Function works fine with archaic ungulate data but will need some further work so that it can handle subgenera and subspecies.