First Update in a While
Updating code in dat and src files as thing have been updated for final run for dissertation.  Updating in part to have backup in case new code form Jon breaks things.  I have moved some code around but that is mostly the scripts folder which is not backed up on github.