Made minor changes to functions
Made comment in oneGenusMat to note that some genera entires also have a subgenus attached so this subgenus needs to be stripped prior to aggregation.  Added some objects to getCorrelationPlots to provide some controls over how the correlation matrices are plotted.  Namely how significance is defined and how that is shown in plot (insig values are blank or sig values given astrix.