Update src_ecologyAnalysisFns.R
Modified code for the Hadley analysis to work with countCube data that is in long format rather than wide.  Did this by swapping around how dimensions (dim()[x]) and indexing between rows and columns so that body sizes are the columns and intervals were the rows.  Having some issues with the sapply() on line 394.  thisCounts is a single element matrix but it the sapply is not treating it as such.  If you go into getLikelihoodOneSetIntBreaks manually (this.count.set = thisCounts[[1]]) the function runs fine.  May revert to older version of this code after this upload to see if transposing countCube to be wide format will fix this (doubtful).