Update src_evanproposal.R
Re-added code (traitHadley and taxHadley) to interface with Jon's Hadley code.  Mainly this get the median species richness from countCube (pseudoreps) and performs Hadley over the medan and all pseudoreplicates and saves the results.  Going to looking into merging the two functions as the taxonomic setup can be done outside this function.