Update NOW_loc_dates.csv
Noticed that a fair number of entries that had their max and min ages swapped for some reason.  This did not affect all entries so there had been a mix of correct and incorrect.  Made a quick line code that parsed for entries with max < min and then swapped them.