Update src_evanproposal.R
Added a few lines of code to correlation function to allow for a bonferroni correction of the p-values.  This code can be toggle off using FALSE or on using bonferroni.  An override for the number of tests was also included with the default being calculated from the product of categories of prop1 and prop 2.