Added Plotting Functionality for Chapter 2 Figures
Added some arguments to make shoulder plots and within regime distributions.  Mainly added the ability to toggle on/off more options for plotting.  Also made a function to dynamically generate vectors that will be passed into the mat argument for layout.  this was done in order to plot shoulder plots alongside the distribution box or bar plots.